ExtractVariable: extract subexpression body
This handles extraction of the body of the subexpression, but the
parameters are not properly extracted yet - they are taken from the
whole expression (the expression surrounding the subexpression).
The extraction of the source cody being an subexpression is taken from
the ExtractVariable.cpp (exact copy). We need to tweak it a little, to
support extraction of the parameters too.